pyproject.toml
[project]
name = "gai"
version = "0.1.9"
description = "Google Gemini prompting script with flexible CLI, templating, and configuration"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Nicolas Girard" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "google-genai>=1.50.1,<2.0.0",
    "tomli>=2.3.0; python_version < '3.11'",
    "jinja2>=3.1.6,<4.0.0",
]

[project.scripts]
gai = "gai.__main__:main"

[project.optional-dependencies]
dev = [
    "ruff>=0.14.5",
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatchling.build.targets.wheel]
packages = ["src/gai"]

[tool.hatchling.build.targets.sdist]
include = ["src/gai"]

[tool.uv]
dev-dependencies = [
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501", # line too long, handled by formatter
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--cov=gai",
    "--cov-report=term-missing",
]
